requirements.txt
streamlit
pandas
requests
beautifulsoup4
selenium
undetected-chromedriver
python-dotenv
wappalyzer
pyhunter
python-dotenv
requests
tqdm